requirements.txt
BigQuery-Python==1.8.0
luigi==2.3.2
slackweb==1.0.5
requests==2.10.0
sqlalchemy==1.0.11
google-api-python-client==1.5.2
csvkit==0.9.1
xmltodict==0.9.2
retrying==1.3.3
scrapelib==1.0.0
beautifulsoup4==4.4.1
html5lib==1.0b8
google-cloud-datastore==0.21.0
boto==2.41.0
Jinja2==2.8
google-cloud-dataflow==0.4.3
luigi-td==0.6.6
